+++ src/main/scala/wdlTools/exec/TaskContext.scala
@@ -27,7 +27,7 @@ case class TaskContext(task: Task,
     // just allow them to be localized on demand (for example, if they're required to
     // evaluate an output value expression).
     WdlValueBindings(if (hasCommand) {
-      val disambiguator = SafeLocalizationDisambiguator(hostEvaluator.paths.getRootDir(true))
+      val disambiguator = SafeLocalizationDisambiguator(hostEvaluator.paths.getRootDir(true))()
       // TODO: put localization behind a trait so we can swap in e.g. a parallelized implementation
       bindings.map {
         case (name, V_File(uri)) =>