+++ RELEASE_NOTES.md
@@ -1,5 +1,14 @@
 # Change log
 
+## 0.15.0 (2021-08-16)
+
+* No longer unwraps optional type for an input with a default value
+* Automatically changes struct declaration with map literal assignment to an object literal 
+* Adds `Runtime.returnCodes` accessor
+* Fixes formatting of single-line expressions in code generator/formatter
+* Fixes default memory value for WDL `development`
+* Always allow coercion to `T_Any`
+
 ## 0.14.3 (2021-07-06)
 
 * Code generator/formatter handles empty blocks